+++ NEWS.md
@@ -8,6 +8,13 @@ Cite all releases with doi [10.5281/zenodo.5475443](https://doi.org/10.5281/zeno
 which always resolves to the latest release.
 
 
+
+* New function `write_mats_to_excel()` writes
+  matrices in a column of a `matsindf` data frame
+  to an Excel file, one matrix per tab.
+  
+
+
 ## matsindf 0.4.10 (2025-05-24) [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.15520543.svg)](https://doi.org/10.5281/zenodo.15520543)
 
 * Updated R dependency to >= 4.1.0 due to using the new pipe.